Extract inventory and owned-item routes
